+++ src/constants.ts
@@ -1,7 +1,10 @@
 import {Decoration} from '@codemirror/view';
 import {wmf} from '@bhsd/common';
+import type {Completion} from '@codemirror/autocomplete';
 
 export const baseData: Record<'CDN', string | undefined> = {CDN: undefined},
+	extData: Record<string, Set<string>> = {},
+	extCompletion: Record<string, Completion[]> = {},
 	hoverSelector = '.cm-tooltip-hover-mw',
 	diagnosticSelector = '.cm-diagnosticText-clickable',
 	panelSelector = '.cm-panel',