+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 <!-- markdownlint-disable first-line-h1 -->
 ## 3.12.3
 
-*2026-03-27*
+*2026-03-28*
 
 **Added**
 
-- Content inside `<math>`, `<chem>` and `<ce>` tags in the MediaWiki mode is now syntax highlighted
-- [LilyPond](https://lilypond.org/) syntax highlighting for content inside `<score>` tags in the MediaWiki mode
+- TeX math syntax highlighting and [autocompletion](./README.md#autocompletion) for content inside `<math>`, `<chem>` and `<ce>` tags in the MediaWiki mode
+- [LilyPond](https://lilypond.org/) syntax highlighting and [autocompletion](./README.md#autocompletion) for content inside `<score>` tags in the MediaWiki mode
 
 ## 3.12.2